+++ library/Test/Tasty/Config.hs
@@ -33,6 +33,7 @@ data Config = Config
   , ignoredModules      :: [FilePath]        -- ^ <<<DEPRECATED>>>: Ignored modules by full name.
   , tastyIngredients    :: [Ingredient]      -- ^ Tasty ingredients to use.
   , tastyOptions        :: [String]          -- ^ Options passed to tasty
+  , inplace             :: Bool              -- ^ Whether the source file should be modified in-place.
   , noModuleSuffix      :: Bool              -- ^ <<<DEPRECATED>>>: suffix and look in all modules.
   , debug               :: Bool              -- ^ Debug the generated module.
   , treeDisplay         :: Bool              -- ^ Tree display for the test results table.
@@ -40,7 +41,7 @@ data Config = Config
 
 -- | The default configuration
 defaultConfig :: Config
-defaultConfig = Config Nothing Nothing Nothing Nothing [] [] [] False False False
+defaultConfig = Config Nothing Nothing Nothing Nothing [] [] [] False False False False
 
 -- | Deprecation message for old `--[no-]module-suffix` option.
 moduleSuffixDeprecationMessage :: String
@@ -97,6 +98,9 @@ options =
   , Option [] ["ingredient"]
       (ReqArg (\s c -> c {tastyIngredients = s : tastyIngredients c}) "INGREDIENT")
       "Qualified tasty ingredient name"
+  , Option [] ["inplace"]
+      (NoArg $ \c -> c {inplace = True})
+      "Whether the source file should be modified in-place"
   , Option [] ["no-module-suffix"]
       (NoArg $ \c -> c {noModuleSuffix = True})
       "<<<DEPRECATED>>>: Ignore test module suffix and import them all"